Create ragged_ops.h
