+++ k2/csrc/array_ops_test.cu
@@ -23,6 +23,8 @@
 #include "k2/csrc/array_ops.h"
 #include "k2/csrc/context.h"
 #include "k2/csrc/math.h"
+#include "k2/csrc/ragged.h"
+#include "k2/csrc/ragged_ops.h"
 #include "k2/csrc/timer.h"
 
 namespace k2 {
