+++ k2/csrc/utils_test.cu
@@ -21,6 +21,7 @@
 #include "k2/csrc/array.h"
 #include "k2/csrc/math.h"
 #include "k2/csrc/ragged.h"
+#include "k2/csrc/ragged_ops.h"
 #include "k2/csrc/utils.h"
 
 namespace k2 {
